+++ Library/Drivers/ADC/adc.h
@@ -1,9 +1,8 @@
 /*
  * adc.h
  *
- * Analog to digital converter library
+ * ADC interrupt and configuration functions
  * 
- * Copyright (c) Carnegie Mellon Racing 2016
  */ 
 
 #ifndef ADC_H_
@@ -14,7 +13,7 @@
 // ADC channel metadata
 typedef struct ADCMeta_t {
 	uint8_t channelNumber;
-	uint8_t currentValue;
+	uint16_t currentValue;
 } ADCMeta;
 
 // Channel data storage
@@ -22,8 +21,8 @@ static ADCMeta adcChannels[NUM_ADC_CH];
 
 // Functions
 void initADC(void);
-uint8_t readADC(uint8_t ch);
+uint16_t readADC(uint8_t ch);
 void updateADC(ADCChannel ch);
-uint8_t adcVal(ADCChannel ch);
+uint16_t adcVal(ADCChannel ch);
 
 #endif /* ADC_H_ */
